backend/requirement.txt
fastapi
pydantic-settings
langchain
langchain-community
langchain-ollama
langchain-qdrant
qdrant-client
pypdf
uvicorn
python-multipart